refactor: changed to snake case
