+++ app/infra/spreadsheet/device_repository.py
@@ -1,6 +1,6 @@
-from typing import List, Optional
+from typing import List
 
-from app.domain.models.bluetooth import BluetoothDevice
+from app.domain.models.device import BluetoothDevice
 from app.domain.repository.device_repository import AbstractDeviceRepository
 from app.infra.spreadsheet.spreadsheet_user_entity import SpreadSheetUserEntity
 from app.infra.spreadsheet.spreadsheet_util import SpreadSheetUtil
@@ -16,17 +16,5 @@ class SpreadSheetDeviceRepository(AbstractDeviceRepository):
             user = SpreadSheetUserEntity.from_csv(values)
             if user is None:
                 continue
-            devices.append(user.toBluetoothDevice())
+            devices.append(user.to_bluetooth_device())
         return devices
-
-    async def find_by_user_id(self, user_id: str) -> Optional[BluetoothDevice]:
-        row = await self.spreadsheet_util.get_row_number_of(user_id, 1)
-        if row == -1:
-            return None
-
-        values = await self.spreadsheet_util.get_row(row)
-        device = SpreadSheetUserEntity.from_csv(values)
-        if device is None:
-            return None
-
-        return device.toBluetoothDevice()
